requirements.txt
django
gunicorn
whitenoise
requests
dotenv
pyjwt